pyproject.toml
[tool.poetry]
name = "django-drone-destination"
version = "0.1.0"
description = ""
authors = ["saizad <mrprasla@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
requests = "^2.31.0"
django-cors-headers = "^4.2.0"
markdown = "^3.4.4"
django-filter = "^23.3"
psycopg2-binary = "^2.9.7"
drf-yasg = "^1.21.7"
pillow = "^10.0.1"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"
redis = "^5.0.0"
django-celery-results = "^2.5.1"
daphne = "^4.0.0"
mixer = "^7.2.2"
django-split-settings = "^1.2.0"
django-environ = "^0.11.2"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
